backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
better-auth==0.0.1b11
asyncpg==0.31.0
qdrant-client==1.11.1
cohere==5.9.3
psycopg[binary]==3.2.3
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
cryptography>=41.0.7
